# openapi.yml
openapi: "3.0.3"
info:
  title: "MyLittleShopping API"
  version: 0.0.1
paths:
...
  /orderProducts:
    get:
      description: "Get orders with separate logic for sellers and buyers"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/orderProduct"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      security:
        - bearerAuth: []

    post:
      description: "Allow the app to create an order when a user validate his cart"
      responses:
        201:
          description: "OrderProduct created"
          content:
            text/plain:
              schema:
                type: string
        400:
          description: "Bad Json Format"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                address1:
                  type: string
                address2:
                  type: string
              required:
                - address1
                - address2

    put:
      description: "Modify a product from the cart"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                items:
                  allOf:
                    - $ref: "#/components/schemas/orderProduct"
        400:
          description: "Bad Json Format"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"

      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                ownerId:
                  type: integer
              required:
                - ownerId
  /syncOrder:
    get:
      description: "Refresh front for caching data"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/orderProduct"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    orderProduct:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        owner_id:
          type: "integer"
          format: "int64"
        product_id:
          type: "integer"
          format: "int64"
        address_id:
          type: "integer"
          format: "int64"
        quantity:
          type: "integer"
          format: "int64"
        created_at:
          type: string
          format: date
        updated_at:
          type: string
          format: date
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT